requirements.txt
beautifulsoup4
requests
yfinance
torch
google-generativeai
python-dotenv